Added BottomSheetDialog for exercise info inserting

+++ app/src/main/java/com/example/oasis/model/Exercise.kt
@@ -1,3 +1,3 @@
 package com.example.oasis.model
 
-data class Exercise(val name: String, val bestResult: Double, var count: Int = 0)
+data class Exercise(val name: String, val bestResult: Int, var count: Int = 0)
