+++ app/src/main/java/com/example/oasis/model/User.kt
@@ -1,5 +1,5 @@
 package com.example.oasis.model
 
-data class User(var name: String, var email: String, val bestResults: Map<String, Double>)
+data class User(var name: String, var email: String, val bestResults: Map<String, Int>)
